Scheuler/back-end/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-decouple==3.8
pytz==2024.1
requests==2.32.3
schedule==1.2.2
scheduler==0.8.7
six==1.16.0
sqlparse==0.5.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
